requirements.txt
# RAGシステムに必要なライブラリ
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-community==0.0.13
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.0
streamlit==1.29.0
sentence-transformers==2.2.2
